requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
pysdl2==0.9.7
    # via -r requirements.in
